refactor(analysis): lift the selection quality pass out of the pipeline
smart_pipeline.py crossed the hard 1000-line gate and left main red. Two
PRs each passed their own gate and together summed past it — neither run
could see the other's lines, which is the failure mode the soft warning at
800 exists to prevent.

The cohesion boundary is the quality pass: selection decides what a memory
is made of, and this decides whether that cut is worth shipping. It moves
whole into a SelectionQuality service with constructor injection —
stabilize, verify, review and final_review_drop, plus judge_offenders and
spare_last_voices, which the judge tests exercise directly.

_UNUSABLE_SHARE_OF_FLOOR travels with spare_last_voices, its only reader.
_verify_attempted stops being pipeline state and belongs to the pass that
owns the memory of what it already tried.

The three dataclasses stay where they are and arrive under TYPE_CHECKING,
so the dependency points one way. That leaves one runtime construction in
the verify loop, which becomes dataclasses.replace(u, analyzed=True) —
identical for a five-field dataclass, and it keeps the import from closing
the loop. Moving the dataclasses instead would have rewritten 35 import
sites.

Pure move: no behaviour change. 1012 -> 642 lines.
